+++ src/Routes.js
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { Route, Switch } from "react-router-dom";
 import UserProfile from "./components/UserProfile";
-// import SignUp from "./components/SignUp";
 import EditUser from "./components/EditUser";
 //import SingleWorkout from "./components/SingleWorkout"
 import { CreateWorkout, SpotifyLogin, SelectPlaylist } from "./components";
@@ -10,7 +9,7 @@ import SignIn from "./components/SignIn";
 import { getAuth, onAuthStateChanged } from "firebase/auth";
 import { fetchLoginUser } from "./store";
 import { useDispatch, useSelector } from "react-redux";
-import SignUpTest from "./components/SignUpTest";
+import SignUp from "./components/SignUp";
 import OptionalSignUp from "./components/OptionalSignUp";
 
 const Routes = () => {
@@ -23,11 +22,17 @@ const Routes = () => {
 
   const dispatch = useDispatch();
 
+  const [isLoading, setLoading] = useState(true);
+
   useEffect(() => {
-    if (!authUser) {
+    if (!authUser.uid || isLoading) {
       dispatch(fetchLoginUser());
     }
-  }, [dispatch, authUser]);
+
+    return () => {
+      setLoading(false);
+    };
+  }, [dispatch, authUser.uid, isLoading]);
 
   // console.log(authUser);
 
@@ -35,6 +40,7 @@ const Routes = () => {
     <div>
       {user ? (
         <Switch>
+          <Route exact path="/optionalsignup" component={OptionalSignUp} />
           <Route exact path="/createworkout" component={CreateWorkout} />
           <Route path="/spotifylogin" component={SpotifyLogin} />
           <Route path="/spotifyhome" component={SelectPlaylist} />
@@ -45,11 +51,10 @@ const Routes = () => {
       ) : (
         <Switch>
           <Route exact path="/signin" component={SignIn} />
-          <Route exact path="/signuptest" component={SignUpTest} />
+          <Route exact path="/signup" component={SignUp} />
           <Route exact path="/" component={SignIn} />
         </Switch>
       )}
-      <Route path="/optionalsignup" component={OptionalSignUp} />
     </div>
   );
 };
